refactor: Extract comp from CategoriesPage
